requirements.txt
mss==10.0.0
PyAutoGUI==0.9.54
pywin32==308